+++ SecurityMine/Controllers/CustomController.cs
@@ -462,239 +462,5 @@ namespace SecurityMine.Controllers
             }
         }
 
-        public ActionResult MedicineSorting(string SortParameter)
-        {
-            AppIdentityDbContext context = new AppIdentityDbContext();
-            AddMedicineValidation dispobj;
-            List<AddMedicineValidation> list = new List<AddMedicineValidation>();
-
-            if (SortParameter.Equals("StockAsc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby s.Stock
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Stock(Low To High)";
-            }
-            else if (SortParameter.Equals("StockDesc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby s.Stock descending
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Stock(High To Low)";
-            }
-            else if (SortParameter.Equals("PriceAsc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.Price
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Price(Low To High)";
-            }
-            else if (SortParameter.Equals("PriceDesc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.Price descending
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Price(High To Low)";
-            }
-            else if (SortParameter.Equals("ExpiryAsc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.Expiry
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Expiry(Early First)";
-            }
-            else if (SortParameter.Equals("ExpiryDesc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.Expiry descending
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Expiry(Late First)";
-            }
-            else if (SortParameter.Equals("MedicineNameAsc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.MedicineName
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Medicine Name(A-Z)";
-            }
-            else if (SortParameter.Equals("MedicineNameDesc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.MedicineName descending
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Medicine Name(Z-A)";
-            }
-            else if (SortParameter.Equals("MedicineTypeAsc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.MedicineType
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Medicine Type(A-Z)";
-            }
-            else if (SortParameter.Equals("MedicineTypeDesc"))
-            {
-                var result = (from m in context.Medicines
-                              join s in context.StoreManagements
-                              on m.MedicineId equals s.MedicineId
-                              orderby m.MedicineType descending
-                              select new { m.MedicineName, m.MedicineType, m.Expiry, m.Price, s.Stock, s.UserId }).ToList();
-
-                //ViewBag.Data = result;
-                foreach (var r in result)
-                {
-                    dispobj = new AddMedicineValidation();
-                    dispobj.MedicineName = r.MedicineName;
-                    dispobj.MedicineType = r.MedicineType;
-                    dispobj.Expiry = r.Expiry;
-                    dispobj.Price = r.Price;
-                    dispobj.Stock = r.Stock;
-
-                    list.Add(dispobj);
-                }
-                ViewBag.SortingDoneBy = "Medicine Type(Z-A)";
-            }
-
-            
-            return View(list);
-
-
-
-        }
-
     }
 }
