+++ SecurityMine/SecurityMine.csproj
@@ -158,6 +158,7 @@
     <Compile Include="Controllers\CustomController.cs" />
     <Compile Include="Controllers\FileManageController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
+    <Compile Include="Controllers\SortAndFilterController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
@@ -236,7 +237,8 @@
     <Content Include="Views\Admin\UserLockoutAction.cshtml" />
     <Content Include="Views\Custom\StoreAddressFromAdmin.cshtml" />
     <Content Include="Views\Admin\UserAddedSuccessfully.cshtml" />
-    <Content Include="Views\Custom\MedicineSorting.cshtml" />
+    <Content Include="Views\SortAndFilter\MedicineSorting.cshtml" />
+    <Content Include="Views\SortAndFilter\FilterMedicine.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="fonts\glyphicons-halflings-regular.woff2" />
@@ -256,6 +258,7 @@
       <DependentUpon>202203230705093_init.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
+  <ItemGroup />
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
     <VSToolsPath Condition="'$(VSToolsPath)' == ''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
